effectsArray2list

Converts an array of posterior samples of multi-trait marker effects to a named list (one for each trait).

effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn, thin)

Arguments

effectsArray
According to BGLR documentation: 3D array, with dim=c(nRow,p,traits), where nRow number of MCMC samples saved, p is the number of predictors and traits is the number of traits. BGLR::Multitrait() writes a binary file to disk when saveEffects=TRUE is specified. It can be read to R with BGLR::readBinMatMultitrait().

snpIDs
character vector with labels for the predictors (SNPs), numeric should work too, but untested.

traits
character vector to label the traits.

nIter
number of iterations used for MCMC; used internally only to exclude burn-in samples from computation

burnIn
burnIn for MCMC; used internally only to exclude burn-in samples from computation

thin
thin for MCMC; used internally only to exclude burn-in samples from computation

Value

list of matrices, one matrix per trait, each matrix has `nrow((nIter-burnIn)/thin)` and `ncol(length(snpIDs))`. Each element of the list is named with a string identifying the trait and the colnames of each matrix are labelled with snpIDs.

Examples

effectsArray<-BGLR::readBinMatMultitrait(filename = "mt_effects_ETA_g_beta.bin")

#> Warning: cannot open compressed file 'mt_effects_ETA_g_beta.bin', probable reason 'No such file or directory'

#> Error in gzfile(filename, open = "rb"): cannot open the connection

effectsArray<-effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn, thin)

#> Error in effectsArray2list(effectsArray, snpIDs, traits, nIter, burnIn,     thin): object 'effectsArray' not found